.netlify/functions/line_webhook_handler/requirements.txt
line-bot-sdk
azure-ai-inference
azure-core
